halpr_foreach, hal_iter is history.
