+++ src/Makefile
@@ -494,7 +494,6 @@ HEADERS := \
     hal/lib/hal_group.h \
     hal/lib/hal.h \
     hal/lib/hal_internal.h \
-    hal/lib/hal_iter.h \
     hal/lib/hal_list.h \
     hal/lib/hal_logging.h \
     hal/lib/hal_parport.h \
