+++ src/hal/lib/Submakefile
@@ -10,7 +10,6 @@ INCLUDES += $(HALLIBDIR)
 HALLIBSRCS := $(HALLIBDIR)/hal_lib.c \
 	$(HALLIBDIR)/hal_group.c \
 	$(HALLIBDIR)/hal_ring.c \
-	$(HALLIBDIR)/hal_iter.c \
 	$(HALLIBDIR)/hal_rcomp.c \
 	$(HALLIBDIR)/hal_vtable.c \
 	$(HALLIBDIR)/hal_funct.c \
@@ -96,6 +95,5 @@ hal_lib-objs += hal/lib/hal_misc.o
 hal_lib-objs += hal/lib/hal_instance.o
 hal_lib-objs += hal/lib/hal_object.o
 hal_lib-objs += hal/lib/hal_object_selectors.o
-hal_lib-objs += hal/lib/hal_iter.o
 
 $(RTLIBDIR)/hal_lib$(MODULE_EXT): $(addprefix $(OBJDIR)/,$(hal_lib-objs))
