+++ src/machinetalk/haltalk/haltalk_introspect.cc
@@ -17,67 +17,79 @@
  */
 
 #include "haltalk.hh"
-#include "hal_iter.h"
+#include "hal_logging.h"
 #include "halpb.hh"
 #include "pbutil.hh"
 #include "rtapi_hexdump.h"
 
 #include <google/protobuf/text_format.h>
 
-static int describe_comp_cb(hal_comp_t *comp,  void *arg);
-static int describe_sig_cb(hal_sig_t *sig,  void *arg);
-static int describe_funct_cb(hal_funct_t *funct,  void *arg);
-static int describe_ring_cb(hal_ring_t *ring,  void *arg);
-static int describe_thread_cb(hal_thread_t *thread,  void *arg);
-static int describe_group_cb(hal_group_t *group,  void *arg);
+static int describe_toplevel_object_cb(hal_object_ptr o,
+				       foreach_args_t *args);
+
+
+// add HAL objects recursively to a Container
+// reflecting the object hierarchy
+int halg_object2pb(const bool use_hal_mutex,
+		   pb::Container *msg,
+		   const char *name,     // selector or NULL
+		   const int type,       // selector or 0
+		   const int id)         // selector or 0
+{
+    WITH_HAL_MUTEX_IF(use_hal_mutex);
+    foreach_args_t args = {};
+    args.name = (char *)name;
+    args.type = type;
+    args.id = id;
+    args.user_ptr1 =  (void *) msg;
+
+    return halg_foreach(use_hal_mutex, &args, describe_toplevel_object_cb);
+}
 
 // describe the current HAL universe as a protobuf message.
 int
-process_describe(htself_t *self, const std::string &from,  void *socket)
+process_describe(htself_t *self,
+		 const std::string &from,
+		 void *socket)
 {
-    int retval __attribute__((cleanup(halpr_autorelease_mutex)));
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    halpr_foreach_comp(NULL, describe_comp_cb, self);
-    halpr_foreach_sig(NULL, describe_sig_cb, self);
-    halpr_foreach_group(NULL, describe_group_cb, self);
-    halpr_foreach_funct(NULL, describe_funct_cb, self);
-    halpr_foreach_ring(NULL, describe_ring_cb, self);
-    halpr_foreach_thread(NULL, describe_thread_cb, self);
+    WITH_HAL_MUTEX();
+    halg_object2pb(0, &self->tx, NULL, 0, 0);
     return send_pbcontainer(from, self->tx, socket);
 }
 
+
 // describe a HAL group as a protobuf message.
 int
-describe_group(htself_t *self, const char *group, const std::string &from,  void *socket)
+describe_group(htself_t *self,
+	       const char *group,
+	       const std::string &from,
+	       void *socket)
 {
-    int retval __attribute__((cleanup(halpr_autorelease_mutex)));
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    hal_group_t *g = halpr_find_group_by_name(group);
-    if (g == NULL) {
+    WITH_HAL_MUTEX();
+    int ret = halg_object2pb(0, &self->tx, group, HAL_GROUP, 0);
+    if (ret != 1)  {
 	self->tx.set_type(pb::MT_HALRCOMP_ERROR);
 	note_printf(self->tx, "no such group: '%s'", group);
 	return send_pbcontainer(from, self->tx, socket);
     }
-    halpr_foreach_group(group, describe_group_cb, self);
     return send_pbcontainer(from, self->tx, socket);
 }
 
+
 // describe a HAL component as a protobuf message.
 int
-describe_comp(htself_t *self, const char *comp, const std::string &from,  void *socket)
+describe_comp(htself_t *self,
+	      const char *comp,
+	      const std::string &from,
+	      void *socket)
 {
-    int retval __attribute__((cleanup(halpr_autorelease_mutex)));
-    rtapi_mutex_get(&(hal_data->mutex));
-
-    hal_comp_t *c = halpr_find_comp_by_name(comp);
-    if (c == NULL) {
+    WITH_HAL_MUTEX();
+    int ret = halg_object2pb(0, &self->tx, comp, HAL_COMPONENT, 0);
+    if (ret != 1)  {
 	self->tx.set_type(pb::MT_HALRCOMP_ERROR);
 	note_printf(self->tx, "no such component: '%s'", comp);
 	return send_pbcontainer(from, self->tx, socket);
     }
-    halpr_foreach_comp(comp, describe_comp_cb, self);
     return send_pbcontainer(from, self->tx, socket);
 }
 
@@ -91,52 +103,48 @@ int describe_parameters(htself_t *self)
     return 0;
 }
 
+
 // ----- end of public functions ---
 
-static int describe_comp_cb(hal_comp_t *comp,  void *arg)
+static int describe_toplevel_object_cb(hal_object_ptr o,
+				       foreach_args_t *args)
 {
-    htself_t *self = (htself_t *) arg;
-    pb::Component *c = self->tx.add_comp();
-    halpr_describe_component(comp, c);
-    return 0;
+    if (!hh_is_toplevel(hh_get_type(o.hdr)))
+	return 0;
+
+    pb::Container *msg = (pb::Container *) args->user_ptr1;
+
+    switch (hh_get_type(o.hdr)) {
+
+    case HAL_SIGNAL:
+	return halpr_describe_signal(o.sig, msg->add_signal());
+	break;
+
+    case HAL_THREAD:
+	return halpr_describe_thread(o.thread, msg->add_thread());
+	break;
+
+    case HAL_COMPONENT:
+	return halpr_describe_component(o.comp,msg->add_comp());
+	break;
+
+    case HAL_VTABLE:
+	// TBD
+	break;
+
+    case HAL_RING:
+	return halpr_describe_ring(o.ring, msg->add_ring());
+	break;
+
+    case HAL_GROUP:
+	return halpr_describe_group(o.group, msg->add_group());
+	break;
+
+    default:
+	HALBUG("%s object %s  not supported",
+	       hh_get_typestr(o.hdr),
+	       hh_get_name(o.hdr));
+    }
+    return 0; // continue
 }
 
-static int describe_sig_cb(hal_sig_t *sig,  void *arg)
-{
-    htself_t *self = (htself_t *) arg;
-    pb::Signal *s = self->tx.add_signal();
-    halpr_describe_signal(sig, s);
-    return 0;
-}
-
-static int describe_group_cb(hal_group_t *g,  void *arg)
-{
-    htself_t *self = (htself_t *) arg;
-    pb::Group *pbgroup = self->tx.add_group();
-    halpr_describe_group(g, pbgroup);
-    return 0;
-}
-
-static int describe_funct_cb(hal_funct_t *funct,  void *arg)
-{
-    htself_t *self = (htself_t *) arg;
-    pb::Function *f = self->tx.add_function();
-    halpr_describe_funct(funct, f);
-    return 0;
-}
-
-static int describe_ring_cb(hal_ring_t *ring,  void *arg)
-{
-    htself_t *self = (htself_t *) arg;
-    pb::Ring *r = self->tx.add_ring();
-    halpr_describe_ring(ring, r);
-    return 0;
-}
-
-static int describe_thread_cb(hal_thread_t *thread,  void *arg)
-{
-    htself_t *self = (htself_t *) arg;
-    pb::Thread *t = self->tx.add_thread();
-    halpr_describe_thread(thread, t);
-    return 0;
-}
