+++ src/machinetalk/proto/proto/object.proto
@@ -133,10 +133,13 @@ message Function {
     optional string      name       = 1; //   [(nanopb).max_size = 41];
     optional fixed32     handle     = 2;
     optional fixed32     owner_id   = 3;
+
     optional fixed32     users      = 4;
     optional fixed32     runtime    = 5;
     optional fixed32     maxtime    = 6;
     optional bool        reentrant  = 7;
+    optional HalFunctType type = 8;
+    optional bool        maxtime_increased  = 9;
 }
 
 message Thread {
@@ -174,6 +177,7 @@ message Component {
 
     repeated Pin         pin            = 16;
     repeated Param       param          = 17;
+    repeated Function    funct          = 19;
 
     // do not create the component on MT_HALRCOMP_BIND if the component does not exist
     optional bool        no_create      = 18;
@@ -249,13 +253,13 @@ message Vtable {
     option (nanopb_msgopt).msgid = 714;
 
     optional string       name          = 1; //  [(nanopb).max_size = 41];
-    optional fixed32      context       = 2;
-    optional sfixed32     comp_id       = 4;
-    optional sfixed32     instance_id   = 5;
-    optional fixed32      handle        = 6;
-    optional sfixed32     refcount      = 7;
-    optional sfixed32     version       = 8;
-    optional fixed64      vtable        = 9;
+    optional fixed32      handle        = 2;
+    optional sfixed32     owner_id      = 3;
+
+    optional fixed32      context       = 5;
+    optional sfixed32     refcount      = 6;
+    optional sfixed32     version       = 7;
+    optional fixed64      vtable        = 8;
 }
 
 message Inst {
@@ -267,5 +271,10 @@ message Inst {
     optional fixed32      inst_id       = 3;
     optional fixed32      inst_size     = 4;
     optional fixed64      vtable        = 5;
+
+    repeated Pin         pin            = 16;
+    repeated Param       param          = 17;
+    repeated Function    funct          = 19;
+
 }
